use APIv3 for buildpacks at the client
