+++ main/cloudfoundry_client/main/main.py
@@ -189,8 +189,9 @@ def main():
                       filter_list_parameters=['organization_guid', 'service_plan_guid'], name_property=None,
                       allow_retrieve_by_name=False, allow_creation=True, allow_deletion=True),
         CommandDomain(display_name='Buildpacks', entity_name='buildpack',
-                      filter_list_parameters=[], allow_retrieve_by_name=False,
-                      allow_creation=False, allow_deletion=False),
+                      api_version='v3',
+                      filter_list_parameters=[], allow_retrieve_by_name=True,
+                      allow_creation=True, allow_deletion=True),
         CommandDomain(display_name='Routes', entity_name='route', name_property='host', filter_list_parameters=[]),
         TaskCommandDomain()
     ]
