+++ test/v2/test_buildpacks.py
@@ -1,7 +1,5 @@
-import sys
 import unittest
 
-import cloudfoundry_client.main.main as main
 from abstract_test_case import AbstractTestCase
 from cloudfoundry_client.imported import OK, reduce
 from fake_requests import mock_response
@@ -45,25 +43,3 @@ class TestBuildpacks(unittest.TestCase, AbstractTestCase):
         self.client.put.assert_called_with(self.client.put.return_value.url,
                                            json=dict(enabled=False))
         self.assertIsNotNone(result)
-
-    @patch.object(sys, 'argv', ['main', 'list_buildpacks'])
-    def test_main_list_buildpacks(self):
-        with patch('cloudfoundry_client.main.main.build_client_from_configuration',
-                        new=lambda: self.client):
-            self.client.get.return_value = mock_response('/v2/buildpacks',
-                                                         OK,
-                                                         None,
-                                                         'v2', 'buildpacks', 'GET_response.json')
-            main.main()
-            self.client.get.assert_called_with(self.client.get.return_value.url)
-
-    @patch.object(sys, 'argv', ['main', 'get_buildpack', '6e72c33b-dff0-4020-8603-bcd8a4eb05e4'])
-    def test_main_get_buildpack(self):
-        with patch('cloudfoundry_client.main.main.build_client_from_configuration',
-                        new=lambda: self.client):
-            self.client.get.return_value = mock_response('/v2/buildpacks/6e72c33b-dff0-4020-8603-bcd8a4eb05e4',
-                                                         OK,
-                                                         None,
-                                                         'v2', 'buildpacks', 'GET_{id}_response.json')
-            main.main()
-            self.client.get.assert_called_with(self.client.get.return_value.url)
